requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
mlflow==2.9.2
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
pylint==3.0.3
jupyter==1.0.0
ipykernel==6.27.1
tensorflow==2.14.0
pillow==10.1.0
opencv-python==4.9.1.74
dvc==2.28.0